Move class RouteCollection into Foundation\Kernel namespace
