+++ packages/framework/src/Foundation/Concerns/BootsHydeKernel.php
@@ -6,7 +6,7 @@ namespace Hyde\Foundation\Concerns;
 
 use Hyde\Foundation\Kernel\FileCollection;
 use Hyde\Foundation\Kernel\PageCollection;
-use Hyde\Foundation\RouteCollection;
+use Hyde\Foundation\Kernel\RouteCollection;
 
 /**
  * @internal Single-use trait for the HydeKernel class.
