+++ packages/framework/src/Foundation/Facades/Router.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace Hyde\Foundation\Facades;
 
 use Hyde\Foundation\HydeKernel;
-use Hyde\Foundation\RouteCollection;
+use Hyde\Foundation\Kernel\RouteCollection;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @mixin \Hyde\Foundation\RouteCollection
+ * @mixin \Hyde\Foundation\Kernel\RouteCollection
  */
 class Router extends Facade
 {
