+++ packages/framework/tests/Feature/HydeExtensionFeatureTest.php
@@ -8,7 +8,7 @@ use Hyde\Foundation\Concerns\HydeExtension;
 use Hyde\Foundation\HydeKernel;
 use Hyde\Foundation\Kernel\FileCollection;
 use Hyde\Foundation\Kernel\PageCollection;
-use Hyde\Foundation\RouteCollection;
+use Hyde\Foundation\Kernel\RouteCollection;
 use Hyde\Hyde;
 use Hyde\Pages\Concerns\HydePage;
 use Hyde\Testing\TestCase;
@@ -20,7 +20,7 @@ use function func_get_args;
  * @covers \Hyde\Foundation\HydeKernel
  * @covers \Hyde\Foundation\Kernel\FileCollection
  * @covers \Hyde\Foundation\Kernel\PageCollection
- * @covers \Hyde\Foundation\RouteCollection
+ * @covers \Hyde\Foundation\Kernel\RouteCollection
  */
 class HydeExtensionFeatureTest extends TestCase
 {
